Finalized scheme for selecting FactoryQoSLib settings. Tested for Component.square_pub_sub.
Note that for <participant_factory_qos base_name="..."> to work,
currently one must override some settings. Added a placeholder:

	<logging>
		<verbosity>ERROR</verbosity>
	</logging>
				
since verbosity is likely what users will want to configure. The
output can be redirected to a file by selecting

	base_name="User.Factory.Logfile"

The FactoryQoSLib offers the selection of choices, and it can 
be expanded further.